Migrate some calls to FHIR and GraphQL (WIP)
The new implementation relies on OkHttp and Apollo.
OkHttp was added because the current implementation relies on Apache HTTP
client, which is not supported anymore since Android 6 (API 23).
A version supports Android 23+, but Apache does not officially maintain it.
Also, OkHttp is the current standard and comes with lots of benefits like,
among other things, the logger, automatic support for GZip, and a strong
security focus.
Apollo is in its version 2 as version 3+ requires Kotlin Coroutines.
Also, the plan is that some new endpoints or extensions could be added to the
FHIR API to remove the dependency on GraphQL.

Everything is hidden behind use cases to impact as less as possible the
current code base.

Also, migrate to AndroidX because of Apollo's dependencies. The versions
of the AndroidX have been kept back to avoid conflicts and upgrades.